pyproject.toml
[project]
name = "nl2sql-eval"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "boto3==1.34.0",
    "pandas==2.1.4",
    "pillow==10.2.0",
    "plotly==5.18.0",
    "python-dotenv>=1.0.1",
    "streamlit==1.31.0",
]
